+++ server/routes/lists.js
@@ -11,8 +11,6 @@ const db = require("../models");
 const ListProductModel = require("../models/listProductModel");
 const UserList = require("../models/userListModel");
 const ProductModel = require("../models/productModel");
-const upload = require("../services/imageUpload");
-const uploadFile = upload.single("image");
 
 const giveServerError = (res) =>
   res.status(500).send({ errors: [{ msg: "Server error" }] });
@@ -158,47 +156,11 @@ const deleteList = async (req, res) => {
   }
 };
 
-const uploadImage = async (req, res) => {
-  const listId = req.params.listId;
-
-  uploadFile(req, res, async (err) => {
-    if (err) {
-      console.log(err);
-      return res.json({
-        success: false,
-        errors: {
-          title: "Image Upload Error",
-          detail: err.message,
-          error: err,
-        },
-      });
-    } else {
-      const imageUrl = req.file.location;
-
-      try {
-        const [updatedList] = await UserList.update(
-          { imageUrl: imageUrl },
-          {
-            where: {
-              id: +listId,
-            },
-          }
-        );
-        res.status(200).json({ updatedList, imageUrl: imageUrl });
-      } catch (error) {
-        res.status(400).send({ error: { msg: "upload fail" } });
-      }
-    }
-  });
-};
-
 router.use(authMiddleware);
 router.get("/", getLists);
 router.get("/:listId", [listIdCheck, validate, getList]);
 router.post("/", [titleCheck, validate, createList]);
 router.put("/:listId", [listIdCheck, titleCheck, validate, changeList]);
 router.delete("/:listId", [listIdCheck, validate, deleteList]);
-
-router.post("/:listId/upload-image", [listIdCheck, validate, uploadImage]);
 
 module.exports = router;
